make it a standalone command
